+++ src/markets/AUDUSD/params.go
@@ -2,12 +2,12 @@ package AUDUSD
 
 import "TradingBot/src/types"
 
-var priceAdjustment float32 = float32(1) / float32(10000)
+var priceAdjustment float64 = float64(1) / float64(10000)
 
 var EMACrossoverLongParams = types.MarketStrategyParams{
 	RiskPercentage: 3,
 
-	StopLossPriceOffset: float64(0 * priceAdjustment),
+	StopLossPriceOffset: 0 * priceAdjustment,
 	CandlesAmountForHorizontalLevel: &types.CandlesAmountForHorizontalLevel{
 		Future: 2,
 		Past:   0,
@@ -17,8 +17,8 @@ var EMACrossoverLongParams = types.MarketStrategyParams{
 	TakeProfitDistance:  200 * priceAdjustment,
 	MinProfit:           10 * priceAdjustment,
 	TrailingStopLoss: &types.TrailingStopLoss{
-		TPDistanceShortForTighterSL: float64(120 * priceAdjustment),
-		SLDistanceWhenTPIsVeryClose: float64(15 * priceAdjustment),
+		TPDistanceShortForTighterSL: 120 * priceAdjustment,
+		SLDistanceWhenTPIsVeryClose: 15 * priceAdjustment,
 	},
 	CandlesAmountWithoutEMAsCrossing: 0,
 	MaxSecondsOpenTrade:              0,
@@ -30,7 +30,7 @@ var EMACrossoverLongParams = types.MarketStrategyParams{
 var EMACrossoverShortParams = types.MarketStrategyParams{
 	RiskPercentage: 3,
 
-	StopLossPriceOffset: float64(200 * priceAdjustment),
+	StopLossPriceOffset: 200 * priceAdjustment,
 	CandlesAmountForHorizontalLevel: &types.CandlesAmountForHorizontalLevel{
 		Future: 25,
 		Past:   40,
@@ -40,8 +40,8 @@ var EMACrossoverShortParams = types.MarketStrategyParams{
 	TakeProfitDistance:  170 * priceAdjustment,
 	MinProfit:           100 * priceAdjustment,
 	TrailingTakeProfit: &types.TrailingTakeProfit{
-		SLDistanceShortForTighterTP: float64(100 * priceAdjustment),
-		TPDistanceWhenSLIsVeryClose: float64(60 * priceAdjustment),
+		SLDistanceShortForTighterTP: 100 * priceAdjustment,
+		TPDistanceWhenSLIsVeryClose: 60 * priceAdjustment,
 	},
 	CandlesAmountWithoutEMAsCrossing: 6,
 	MaxSecondsOpenTrade:              0,
