+++ src/markets/USDCHF/params.go
@@ -2,12 +2,12 @@ package USDCHF
 
 import "TradingBot/src/types"
 
-var priceAdjustment float32 = float32(1) / float32(10000)
+var priceAdjustment float64 = float64(1) / float64(10000)
 
 var EMACrossoverShortParams = types.MarketStrategyParams{
 	RiskPercentage: 3,
 
-	StopLossPriceOffset: float64(25 * priceAdjustment),
+	StopLossPriceOffset: 25 * priceAdjustment,
 	CandlesAmountForHorizontalLevel: &types.CandlesAmountForHorizontalLevel{
 		Future: 25,
 		Past:   10,
@@ -17,8 +17,8 @@ var EMACrossoverShortParams = types.MarketStrategyParams{
 	TakeProfitDistance:  120 * priceAdjustment,
 	MinProfit:           60 * priceAdjustment,
 	TrailingTakeProfit: &types.TrailingTakeProfit{
-		SLDistanceShortForTighterTP: float64(20 * priceAdjustment),
-		TPDistanceWhenSLIsVeryClose: float64(-100 * priceAdjustment),
+		SLDistanceShortForTighterTP: 20 * priceAdjustment,
+		TPDistanceWhenSLIsVeryClose: -100 * priceAdjustment,
 	},
 	CandlesAmountWithoutEMAsCrossing: 10,
 	MaxSecondsOpenTrade:              0,
